requirements.txt
python-telegram-bot
cffi
numpy
polyglot
pyicu
pycld2
morfessor
python-dotenv